[REFACTOR] データを操作するすべてのメソッドを個別のクラスに切り出す
